requirements.txt
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cupy-cuda12x==13.6.0
cycler==0.12.1
egrabber @ file:///C:/Program%20Files/Euresys/eGrabber/python/egrabber-24.01.2.71-py2.py3-none-any.whl#sha256=f0cac1edcf63157b1435f9a1f6e59ce8ca9450acec3138104696c0736018197a
fastrlock==0.8.3
fonttools==4.61.1
kiwisolver==1.4.9
magscope @ git+https://github.com/7jameslondon/MagScope.git@397e87e0eec4a81be2575e5ada181317f1b9172f
magtrack==0.7.6
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pillow==12.1.1
pipython==2.11.0.6
pyparsing==3.3.2
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
reactivex==4.1.0
scipy==1.17.0
six==1.17.0
tifffile==2026.1.28
typing_extensions==4.15.0
zaber_motion==8.2.0
